Distinguish between banks' shared and testing configuration.
