+++ src/Entity/TestSPankkiBank.php
@@ -3,11 +3,11 @@
 namespace Tupas\Entity;
 
 /**
- * Defines Handelsbanken.
+ * Defines the test S-Pankki bank.
  *
  * @package Tupas\Entity
  */
-final class HandelsbankenBank extends BaseBank
+final class TestSPankkiBank extends BaseSPankkiBank
 {
 
   /**
@@ -15,13 +15,12 @@ final class HandelsbankenBank extends BaseBank
    */
     public function __construct()
     {
-        $this->actionUrl = 'https://tunnistepalvelu.samlink.fi/TupasTunnistus/SHBtupas.html';
+        parent::__construct();
         $this->certVersion = '0002';
-        $this->receiverId = '1111111111111';
-        $this->receiverKey = '11111111111111111111';
+        $this->receiverId = 'SPANKKITUPAS';
+        $this->receiverKey = 'SPANKKI';
         $this->idType = '02';
         $this->keyVersion = '0001';
         $this->algorithm = '03';
-        $this->bankNumber = 310;
     }
 }
